centralizing to github_iterator
